No se podia meter numero decimal en promedio, ahora ya funciona aumente type number en html create

+++ VelascoCrhystel_PruebaProgreso_1/Models/Velasco.cs
@@ -12,7 +12,7 @@ namespace VelascoCrhystel_PruebaProgreso_1.Models
         public string Nombre { get; set; }
         [Required]
         [Range(0.0,10.0)]
-        public decimal Promedio { get; set; }
+        public float Promedio { get; set; }
         [Display(Name ="Tiene Ganas de Estudiar?")]
         public bool TieneGanasEstudiar { get; set; }
         [DataType(DataType.Date)]
